SQUASH ME: address length field fix in header

+++ sys/include/net/gnrc/ndp/node.h
@@ -41,7 +41,7 @@ extern "C" {
  * @return  -ENOBUFS, if @p l2addr_len was smaller than the resulting @p l2addr
  *          would be long.
  */
-kernel_pid_t gnrc_ndp_node_next_hop_l2addr(uint8_t *l2addr, uint8_t *l2addr_len,
+kernel_pid_t gnrc_ndp_node_next_hop_l2addr(uint8_t *l2addr, uint16_t *l2addr_len,
                                            kernel_pid_t iface, ipv6_addr_t *dst,
                                            gnrc_pktsnip_t *pkt);
